First commit of CMake'd MEA. Warnings as errors now enabled, as are pedantic warnings. These have been fixed up throughout the project, with the exception of external code (ICarousel, Yelp/OAAuth etc), in which case warnings have been suppressed. Buddy: Paul H

+++ ios/ios_src/Menu/View/CustomTableDataProvider.mm
@@ -100,7 +100,7 @@ NSInteger const SubItemCellOpenableMenuArrowTag = 1;
     }
     else
     {
-        ExampleApp::Menu::View::MenuItemModel item = section.GetItemAtIndex(index);
+        ExampleApp::Menu::View::MenuItemModel item = section.GetItemAtIndex(static_cast<int>(index));
         std::string json = item.SerializeJson();
         [self populateCellWithJson :json :cell :isHeader];
     }
@@ -158,7 +158,7 @@ NSInteger const SubItemCellOpenableMenuArrowTag = 1;
     else
     {
         ExampleApp::Menu::View::MenuViewInterop* interop = [m_pView getInterop];
-        interop->HandleItemSelected(indexPath.section, indexPath.row);
+        interop->HandleItemSelected(static_cast<int>(indexPath.section), static_cast<int>(indexPath.row));
     }
 }
